Refactor SmsSender extension to controller structure
