+++ hugashop/crm/Extensions/SmsSender/settings.yaml
@@ -1,3 +1,3 @@
 name: "SMS рассылка"
 description: "Рассылка СМС сообщений по пользователям. Собирает список рассылки по Товарам и Категориям товаров"
-version: 1.6
+version: 1.7
